// src/enums/Weapon.ts
export enum Weapon {
    Shield = 0,
    Fist = 1,
    Wand = 2,
    Staff = 3,
    Dagger = 4,
    Bow = 5,
    Spear = 6,
    Sword = 7,
    Crossbow = 8,
    Mace = 9,
    Axe = 10,
    Flail = 11,
    Hammer = 12,
    Scythe = 13,
    Whip = 14,
    Gun = 15,
    Grenade = 16,
    Bomb = 17,
    Rocket = 18,
    Laser = 19,
}